moving resequence_labels to tools and exposing
